Added "user_date" field to writing model
Since the user will want to know if they met their word count goals for
a given day *within their own timezone*, it makes sense to keep track of
what day the user completed their writing based on their timezone.

This user_date could theoretically be calculated from the time field
already in the model. However, if I did it this way, searching the db
for a given local date would require converting back and forth between
UTC and the user's timezone.  This is fairly messy. When the local date
itself is stored, we only have to do the conversion one way (from UTC to
local date), and not the local date to UTC conversion which is messier.

Also, if the user changes their current timezone, it probably shouldn't
change the past. If they completed their goal for a given date, they
should still get credit for completing that goal even if they move to a
different timezone later on. But if we dynamically calculate based on
their current timezone, and a user changes their timezone, it could
change the way their writings are split between days and wreak havoc on
the report of their past acheivements.