"Reset" goal tracking after change in tz/goal
In general, changes in timezone and target word count could lead to
surprising behavior when viewing past achievements/failures/streaks.
"Resetting" that old information - setting the "streak" length to 0, and
not displaying successes/failures prior to the last reset - conveys to
users that the old information is no longer valid after such changes.